Refactored get_func() as staticmethod
It makes more sense to treat get_func() as a method that isn't tied to a
particular instance of a DefaultEngine class. Instead, pass the class as
a parameter. Also, accept a method name as an argument so that the
method can be used to curry multiple methods, not just get_page_source()